chore(api): migrate campaign participation lifecycle to its bt pounded context
