+++ api/lib/domain/models/CampaignParticipation.js
@@ -1,13 +1,4 @@
 import _ from 'lodash';
-
-import {
-  ArchivedCampaignError,
-  AssessmentNotCompletedError,
-  AlreadySharedCampaignParticipationError,
-  CantImproveCampaignParticipationError,
-  CampaignParticipationDeletedError,
-} from '../errors.js';
-
 import { CampaignParticipationStatuses } from '../../../src/prescription/shared/domain/constants.js';
 
 class CampaignParticipation {
@@ -42,20 +33,6 @@ class CampaignParticipation {
     this.organizationLearnerId = organizationLearnerId;
   }
 
-  static start(campaignParticipation) {
-    const { organizationLearnerId = null } = campaignParticipation;
-    const { isAssessment } = campaignParticipation.campaign;
-    const { STARTED, TO_SHARE } = CampaignParticipationStatuses;
-
-    const status = isAssessment ? STARTED : TO_SHARE;
-
-    return new CampaignParticipation({
-      ...campaignParticipation,
-      status,
-      organizationLearnerId,
-    });
-  }
-
   get isShared() {
     return this.status === CampaignParticipationStatuses.SHARED;
   }
@@ -76,46 +53,10 @@ class CampaignParticipation {
     return _.get(this, 'campaign.id', null);
   }
 
-  share() {
-    this._canBeShared();
-    this.sharedAt = new Date();
-    this.status = CampaignParticipationStatuses.SHARED;
-  }
-
-  improve() {
-    this._canBeImproved();
-    this.status = CampaignParticipationStatuses.STARTED;
-  }
-
   delete(userId) {
     this.deletedAt = new Date();
     this.deletedBy = userId;
   }
-
-  _canBeImproved() {
-    if (this.campaign.isProfilesCollection()) {
-      throw new CantImproveCampaignParticipationError();
-    }
-  }
-
-  _canBeShared() {
-    if (this.isShared) {
-      throw new AlreadySharedCampaignParticipationError();
-    }
-    if (this.campaign.isArchived()) {
-      throw new ArchivedCampaignError('Cannot share results on an archived campaign.');
-    }
-    if (this.isDeleted) {
-      throw new CampaignParticipationDeletedError('Cannot share results on a deleted participation.');
-    }
-    if (this.campaign.isAssessment() && lastAssessmentNotCompleted(this)) {
-      throw new AssessmentNotCompletedError();
-    }
-  }
-}
-
-function lastAssessmentNotCompleted(campaignParticipation) {
-  return !campaignParticipation.lastAssessment || !campaignParticipation.lastAssessment.isCompleted();
 }
 
 export { CampaignParticipation };
