+++ api/src/prescription/campaign-participation/domain/models/CampaignParticipant.js
@@ -1,12 +1,13 @@
 import {
   AlreadyExistingCampaignParticipationError,
   NotEnoughDaysPassedBeforeResetCampaignParticipationError,
-} from '../errors.js';
-import { ForbiddenAccess, EntityValidationError } from '../../../src/shared/domain/errors.js';
+} from '../../../../../lib/domain/errors.js';
+import { ForbiddenAccess, EntityValidationError } from '../../../../shared/domain/errors.js';
 
 import { CampaignParticipation } from './CampaignParticipation.js';
-import { Assessment } from '../../../src/shared/domain/models/Assessment.js';
-import { OrganizationLearner } from './OrganizationLearner.js';
+import { Assessment } from '../../../../shared/domain/models/Assessment.js';
+import { OrganizationLearner } from '../../../../../lib/domain/models/OrganizationLearner.js';
+
 const couldNotJoinCampaignErrorMessage = "Vous n'êtes pas autorisé à rejoindre la campagne";
 const couldNotImproveCampaignErrorMessage = 'Vous ne pouvez pas repasser la campagne';
