+++ api/src/prescription/campaign-participation/domain/usecases/start-campaign-participation.js
@@ -1,5 +1,5 @@
-import { CampaignParticipationStarted } from '../events/CampaignParticipationStarted.js';
-import { KnowledgeElement } from '../models/KnowledgeElement.js';
+import { CampaignParticipationStarted } from '../../../../../lib/domain/events/CampaignParticipationStarted.js';
+import { KnowledgeElement } from '../../../../../lib/domain/models/KnowledgeElement.js';
 
 const startCampaignParticipation = async function ({
   campaignParticipation,
