+++ api/src/prescription/campaign-participation/routes.js
@@ -1,0 +1,5 @@
+import * as learnerParticipations from './application/learner-participation-route.js';
+
+const campaignParticipationsRoutes = [learnerParticipations];
+
+export { campaignParticipationsRoutes };
