+++ api/src/prescription/organization-learner/routes.js
@@ -3,6 +3,6 @@ import * as scoLearnerListRoute from './application/sco-learner-list-route.js';
 import * as learnerListRoute from './application/learner-list-route.js';
 import * as leanerActivityRoute from './application/learner-activity-route.js';
 
-const learnerListRoutes = [supLearnerListRoute, scoLearnerListRoute, learnerListRoute, leanerActivityRoute];
+const organizationLearnerRoutes = [supLearnerListRoute, scoLearnerListRoute, learnerListRoute, leanerActivityRoute];
 
-export { learnerListRoutes };
+export { organizationLearnerRoutes };
