+++ api/tests/integration/domain/usecases/complete-assessment_test.js
@@ -1,7 +1,7 @@
 import { expect, databaseBuilder, knex } from '../../../test-helper.js';
 import { completeAssessment } from '../../../../lib/domain/usecases/complete-assessment.js';
 import * as assessmentRepository from '../../../../src/shared/infrastructure/repositories/assessment-repository.js';
-import * as campaignParticipationRepository from '../../../../lib/infrastructure/repositories/campaign-participation-repository.js';
+import * as campaignParticipationBCRepository from '../../../../src/prescription/campaign-participation/infrastructure/repositories/campaign-participation-repository.js';
 import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
 import { CampaignParticipationStatuses } from '../../../../src/prescription/shared/domain/constants.js';
 
@@ -39,7 +39,7 @@ describe('Integration | Usecase | Complete Assessment', function () {
           await completeAssessment({
             assessmentId,
             domainTransaction,
-            campaignParticipationRepository,
+            campaignParticipationBCRepository,
             assessmentRepository,
           });
