+++ api/tests/integration/infrastructure/repositories/campaign-participant-repository_test.js
@@ -1,599 +1,11 @@
 import { catchErr, databaseBuilder, expect, mockLearningContent } from '../../../test-helper.js';
-import { knex } from '../../../../db/knex-database-connection.js';
 import * as campaignParticipantRepository from '../../../../lib/infrastructure/repositories/campaign-participant-repository.js';
-import { CampaignParticipant } from '../../../../lib/domain/models/CampaignParticipant.js';
-import { CampaignToStartParticipation } from '../../../../lib/domain/models/CampaignToStartParticipation.js';
-import lodash from 'lodash';
-import {
-  AlreadyExistingCampaignParticipationError,
-  OrganizationLearnersCouldNotBeSavedError,
-  NotFoundError,
-} from '../../../../lib/domain/errors.js';
 import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
 
-const { pick } = lodash;
-
-const campaignParticipationDBAttributes = [
-  'id',
-  'campaignId',
-  'userId',
-  'status',
-  'organizationLearnerId',
-  'participantExternalId',
-];
-
-const assessmentAttributes = ['userId', 'method', 'state', 'type', 'courseId', 'isImproving'];
+import { CampaignToStartParticipation } from '../../../../src/prescription/campaign-participation/domain/models/CampaignToStartParticipation.js';
+import { NotFoundError } from '../../../../src/shared/domain/errors.js';
 
 describe('Integration | Infrastructure | Repository | CampaignParticipant', function () {
-  describe('save', function () {
-    let userIdentity;
-
-    beforeEach(async function () {
-      const user = databaseBuilder.factory.buildUser();
-      await databaseBuilder.commit();
-      userIdentity = { id: user.id, firstName: user.firstName, lastName: user.lastName };
-    });
-
-    it('returns campaign participation id', async function () {
-      const campaignParticipant = await makeCampaignParticipant({
-        campaignAttributes: { idPixLabel: null },
-        userIdentity,
-        participantExternalId: null,
-        isRestricted: false,
-      });
-
-      const id = await DomainTransaction.execute(async (domainTransaction) => {
-        return campaignParticipantRepository.save(campaignParticipant, domainTransaction);
-      });
-
-      const [campaignParticipationId] = await knex('campaign-participations').pluck('id');
-
-      expect(id).to.equal(campaignParticipationId);
-    });
-
-    context('when user is disabled and the orga isManagingStudent is set to false', function () {
-      let learnerDisabled, otherLearnerDisabled, campaignParticipant;
-      beforeEach(async function () {
-        const orga = databaseBuilder.factory.buildOrganization({
-          isManagingStudents: false,
-        });
-        learnerDisabled = databaseBuilder.factory.buildOrganizationLearner({
-          userId: userIdentity.id,
-          isDisabled: true,
-          organizationId: orga.id,
-        });
-
-        otherLearnerDisabled = databaseBuilder.factory.buildOrganizationLearner({
-          isDisabled: true,
-          organizationId: orga.id,
-        });
-
-        await databaseBuilder.commit();
-
-        campaignParticipant = await makeCampaignParticipant({
-          campaignAttributes: { organizationId: orga.id, idPixLabel: null },
-          userIdentity,
-          participantExternalId: 'null',
-          isRestricted: orga.isManagingStudents,
-        });
-      });
-
-      it('creates a campaign participation', async function () {
-        // when
-        await DomainTransaction.execute(async (domainTransaction) => {
-          await campaignParticipantRepository.save(campaignParticipant, domainTransaction);
-        });
-
-        const campaignParticipation = await knex('campaign-participations')
-          .select(campaignParticipationDBAttributes)
-          .first();
-
-        expect(campaignParticipation).to.deep.equal(
-          getExpectedCampaignParticipation(campaignParticipation.id, campaignParticipant),
-        );
-      });
-
-      it('enabled only the learner assigned to the campaign participant', async function () {
-        // when
-        await DomainTransaction.execute(async (domainTransaction) => {
-          await campaignParticipantRepository.save(campaignParticipant, domainTransaction);
-        });
-
-        const { isDisabled: expectedEnabledLearner } = await knex('organization-learners')
-          .select('isDisabled')
-          .where('id', learnerDisabled.id)
-          .first();
-
-        const { isDisabled: expectedDisabledLearner } = await knex('organization-learners')
-          .select('isDisabled')
-          .where('id', otherLearnerDisabled.id)
-          .first();
-
-        expect(expectedEnabledLearner).to.be.false;
-        expect(expectedDisabledLearner).to.be.true;
-      });
-    });
-
-    context('when the campaign is profile collection', function () {
-      it('creates a campaign participation', async function () {
-        const campaignParticipant = await makeCampaignParticipant({
-          campaignAttributes: { type: 'PROFILES_COLLECTION', idPixLabel: null },
-          userIdentity,
-          participantExternalId: null,
-          isRestricted: false,
-        });
-
-        await DomainTransaction.execute(async (domainTransaction) => {
-          await campaignParticipantRepository.save(campaignParticipant, domainTransaction);
-        });
-
-        const campaignParticipation = await knex('campaign-participations')
-          .select(campaignParticipationDBAttributes)
-          .first();
-
-        expect(campaignParticipation).to.deep.equal(
-          getExpectedCampaignParticipation(campaignParticipation.id, campaignParticipant),
-        );
-      });
-
-      it('does not create an assessment', async function () {
-        const campaignParticipant = await makeCampaignParticipant({
-          campaignAttributes: { type: 'PROFILES_COLLECTION', idPixLabel: null },
-          userIdentity,
-          participantExternalId: null,
-          isRestricted: false,
-        });
-
-        await DomainTransaction.execute(async (domainTransaction) => {
-          await campaignParticipantRepository.save(campaignParticipant, domainTransaction);
-        });
-
-        const assessments = await knex('assessments');
-
-        expect(assessments).to.be.empty;
-      });
-    });
-
-    context('when the campaign is assessment', function () {
-      it('create a campaign participation and an assessment', async function () {
-        //GIVEN
-        const campaignParticipant = await makeCampaignParticipant({
-          campaignAttributes: {
-            type: 'ASSESSMENT',
-            idPixLabel: null,
-            method: 'SMART_RANDOM',
-          },
-          userIdentity,
-          participantExternalId: null,
-          isRestricted: false,
-        });
-
-        //WHEN
-        await DomainTransaction.execute(async (domainTransaction) => {
-          await campaignParticipantRepository.save(campaignParticipant, domainTransaction);
-        });
-
-        //THEN
-        const campaignParticipation = await knex('campaign-participations')
-          .select(['id', ...campaignParticipationDBAttributes])
-          .first();
-
-        const assessment = await knex('assessments')
-          .select(['campaignParticipationId', ...assessmentAttributes])
-          .first();
-
-        expect(campaignParticipation).to.deep.equal(
-          getExpectedCampaignParticipation(campaignParticipation.id, campaignParticipant),
-        );
-        expect(assessment).to.deep.equal(getExpectedAssessment(campaignParticipation.id, campaignParticipant));
-      });
-    });
-
-    context('when there is already an organization learner', function () {
-      it('create a campaign participation linked to this organization learner', async function () {
-        //GIVEN
-        const campaign = databaseBuilder.factory.buildCampaign({ idPixLabel: null });
-        const organizationLearnerId = databaseBuilder.factory.buildOrganizationLearner().id;
-        await databaseBuilder.commit();
-
-        const campaignToStartParticipation = new CampaignToStartParticipation(campaign);
-        const campaignParticipant = new CampaignParticipant({
-          campaignToStartParticipation,
-          organizationLearnerId,
-          organizationLearner: {
-            id: organizationLearnerId,
-            hasParticipated: false,
-          },
-          userIdentity,
-          previousCampaignParticipationForUser: null,
-        });
-
-        campaignParticipant.start({ participantExternalId: null });
-
-        //WHEN
-        await DomainTransaction.execute(async (domainTransaction) => {
-          await campaignParticipantRepository.save(campaignParticipant, domainTransaction);
-        });
-
-        //THEN
-        const campaignParticipation = await knex('campaign-participations').select('organizationLearnerId').first();
-        expect(campaignParticipation.organizationLearnerId).to.equal(organizationLearnerId);
-      });
-    });
-
-    context('when there is no organization learner linked', function () {
-      it('create a new organization learner', async function () {
-        //GIVEN
-        userIdentity = databaseBuilder.factory.buildUser({ firstName: 'Valentin', lastName: 'Tamare' });
-        const campaign = databaseBuilder.factory.buildCampaign({ idPixLabel: null });
-        await databaseBuilder.commit();
-
-        const campaignToStartParticipation = new CampaignToStartParticipation(campaign);
-        const campaignParticipant = new CampaignParticipant({
-          campaignToStartParticipation,
-          organizationLearner: {
-            id: null,
-            hasParticipated: false,
-          },
-          userIdentity,
-          previousCampaignParticipationForUser: null,
-        });
-
-        campaignParticipant.start({ participantExternalId: null });
-
-        //WHEN
-        await DomainTransaction.execute(async (domainTransaction) => {
-          await campaignParticipantRepository.save(campaignParticipant, domainTransaction);
-        });
-
-        //THEN
-        const organizationLearner = await knex('organization-learners')
-          .select('firstName', 'lastName', 'userId', 'organizationId')
-          .first();
-
-        expect(organizationLearner).to.deep.equal({
-          firstName: userIdentity.firstName,
-          lastName: userIdentity.lastName,
-          organizationId: campaignToStartParticipation.organizationId,
-          userId: userIdentity.id,
-        });
-      });
-
-      it('create a campaign participation linked to the new organization learner', async function () {
-        //GIVEN
-        const campaign = databaseBuilder.factory.buildCampaign({ idPixLabel: null });
-        await databaseBuilder.commit();
-
-        const campaignToStartParticipation = new CampaignToStartParticipation(campaign);
-        const campaignParticipant = new CampaignParticipant({
-          campaignToStartParticipation,
-          organizationLearner: {
-            id: null,
-            hasParticipated: false,
-          },
-          userIdentity,
-          previousCampaignParticipationForUser: null,
-        });
-
-        campaignParticipant.start({ participantExternalId: null });
-
-        //WHEN
-        await DomainTransaction.execute(async (domainTransaction) => {
-          await campaignParticipantRepository.save(campaignParticipant, domainTransaction);
-        });
-
-        //THEN
-        const campaignParticipation = await knex('campaign-participations').select('organizationLearnerId').first();
-        const organizationLearner = await knex('organization-learners').select('id').first();
-        expect(campaignParticipation.organizationLearnerId).to.equal(organizationLearner.id);
-      });
-    });
-
-    context('when there is a previous participation', function () {
-      it('update the previous participation', async function () {
-        //GIVEN
-        const campaign = databaseBuilder.factory.buildCampaign({
-          multipleSendings: true,
-        });
-        const { id: previousCampaignParticipationForUserId } = databaseBuilder.factory.buildCampaignParticipation({
-          userId: userIdentity.id,
-          campaignId: campaign.id,
-          isImproved: false,
-          status: 'SHARED',
-        });
-
-        await databaseBuilder.commit();
-
-        const campaignToStartParticipation = new CampaignToStartParticipation(campaign);
-        const campaignParticipant = new CampaignParticipant({
-          campaignToStartParticipation,
-          userIdentity,
-          previousCampaignParticipationForUser: {
-            id: previousCampaignParticipationForUserId,
-            status: 'SHARED',
-            validatedSkillsCount: 0,
-          },
-          organizationLearner: {
-            id: null,
-            hasParticipated: false,
-          },
-        });
-
-        campaignParticipant.start({ participantExternalId: null });
-
-        //WHEN
-        await DomainTransaction.execute(async (domainTransaction) => {
-          await campaignParticipantRepository.save(campaignParticipant, domainTransaction);
-        });
-
-        //THEN
-        const campaignParticipation = await knex('campaign-participations')
-          .select('isImproved')
-          .where({ id: previousCampaignParticipationForUserId })
-          .first();
-
-        expect(campaignParticipation.isImproved).to.deep.equal(true);
-      });
-
-      it('does not update participation for other user or campaign', async function () {
-        //GIVEN
-        const campaign = databaseBuilder.factory.buildCampaign({
-          idPixLabel: null,
-          multipleSendings: true,
-        });
-        databaseBuilder.factory.buildCampaignParticipation({
-          isImproved: false,
-        });
-        databaseBuilder.factory.buildCampaignParticipation({
-          userId: userIdentity.id,
-          isImproved: false,
-        });
-        databaseBuilder.factory.buildCampaignParticipation({
-          campaignId: campaign.id,
-          isImproved: false,
-        });
-        const { id: previousCampaignParticipationForUserId } = databaseBuilder.factory.buildCampaignParticipation({
-          userId: userIdentity.id,
-          campaignId: campaign.id,
-          isImproved: false,
-        });
-
-        await databaseBuilder.commit();
-
-        const campaignToStartParticipation = new CampaignToStartParticipation(campaign);
-        const campaignParticipant = new CampaignParticipant({
-          campaignToStartParticipation,
-          userIdentity,
-          previousCampaignParticipationForUser: {
-            id: previousCampaignParticipationForUserId,
-            status: 'SHARED',
-            validatedSkillsCount: 0,
-          },
-          organizationLearner: {
-            id: null,
-            hasParticipated: false,
-          },
-        });
-
-        campaignParticipant.start({ participantExternalId: null });
-
-        //WHEN
-        await DomainTransaction.execute(async (domainTransaction) => {
-          await campaignParticipantRepository.save(campaignParticipant, domainTransaction);
-        });
-
-        //THEN
-        const campaignParticipations = await knex('campaign-participations').pluck('id').where({ isImproved: true });
-
-        expect(campaignParticipations).to.deep.equal([previousCampaignParticipationForUserId]);
-      });
-    });
-
-    context('when external id is asked', function () {
-      it('save participant external id', async function () {
-        const campaignParticipant = await makeCampaignParticipant({
-          campaignAttributes: { idPixLabel: 'some external id' },
-          userIdentity,
-          participantExternalId: 'some participant external id',
-          isRestricted: false,
-        });
-
-        await DomainTransaction.execute(async (domainTransaction) => {
-          await campaignParticipantRepository.save(campaignParticipant, domainTransaction);
-        });
-
-        const campaignParticipation = await knex('campaign-participations')
-          .select(campaignParticipationDBAttributes)
-          .first();
-
-        expect(campaignParticipation).to.deep.equal(
-          getExpectedCampaignParticipation(campaignParticipation.id, campaignParticipant),
-        );
-      });
-    });
-
-    context('when there is a deleted organization learner', function () {
-      it('should create new participation with new organization learner', async function () {
-        const { id: organizationId } = databaseBuilder.factory.buildOrganization();
-
-        const { id: deletedOrganizationLearnerId } = databaseBuilder.factory.buildOrganizationLearner({
-          organizationId,
-          deletedAt: new Date(),
-        });
-
-        const { id: campaignId } = databaseBuilder.factory.buildCampaign({
-          organizationId,
-          idPixLabel: null,
-        });
-        databaseBuilder.factory.buildCampaignParticipation({
-          organizationLearnerId: deletedOrganizationLearnerId,
-          deletedAt: new Date(),
-          campaignId,
-        });
-
-        await databaseBuilder.commit();
-
-        const campaignParticipant = await makeCampaignParticipant({
-          campaignAttributes: { idPixLabel: null },
-          userIdentity,
-          participantExternalId: null,
-          isRestricted: false,
-        });
-
-        const id = await DomainTransaction.execute(async (domainTransaction) => {
-          return campaignParticipantRepository.save(campaignParticipant, domainTransaction);
-        });
-
-        const startedParticipation = await knex('campaign-participations').where('id', id).first();
-
-        expect(startedParticipation.organizationLearnerId).not.to.equal(deletedOrganizationLearnerId);
-      });
-    });
-
-    context('when there is an exception', function () {
-      context('when there already is a participation for this campaign', function () {
-        it('throws an exception AlreadyExistingCampaignParticipationError', async function () {
-          //GIVEN
-          const campaign = databaseBuilder.factory.buildCampaign({
-            idPixLabel: null,
-          });
-          databaseBuilder.factory.buildCampaignParticipation({
-            userId: userIdentity.id,
-            campaignId: campaign.id,
-            isImproved: false,
-          });
-
-          await databaseBuilder.commit();
-
-          const campaignToStartParticipation = new CampaignToStartParticipation(campaign);
-          const campaignParticipant = new CampaignParticipant({
-            campaignToStartParticipation,
-            userIdentity,
-            organizationLearner: {
-              id: null,
-              hasParticipated: false,
-            },
-          });
-
-          campaignParticipant.start({ participantExternalId: null });
-
-          //WHEN
-          const error = await catchErr(() => {
-            return DomainTransaction.execute(async (domainTransaction) => {
-              await campaignParticipantRepository.save(campaignParticipant, domainTransaction);
-            });
-          })();
-
-          //THEN
-          expect(error).to.be.an.instanceof(AlreadyExistingCampaignParticipationError);
-          expect(error.message).to.equal(
-            `User ${userIdentity.id} has already a campaign participation with campaign ${campaign.id}`,
-          );
-        });
-      });
-
-      context('when there is another error', function () {
-        it('throws the original exception', async function () {
-          //GIVEN
-          const campaign = databaseBuilder.factory.buildCampaign({
-            idPixLabel: null,
-          });
-
-          await databaseBuilder.commit();
-
-          const campaignToStartParticipation = new CampaignToStartParticipation(campaign);
-          const campaignParticipant = new CampaignParticipant({
-            campaignToStartParticipation,
-            userIdentity: { id: 12, firstName: '', lastName: '' },
-            organizationLearner: {
-              id: null,
-              hasParticipated: false,
-            },
-          });
-
-          campaignParticipant.start({ participantExternalId: null });
-
-          //WHEN
-          const error = await catchErr(() => {
-            return DomainTransaction.execute(async (domainTransaction) => {
-              await campaignParticipantRepository.save(campaignParticipant, domainTransaction);
-            });
-          })();
-
-          //THEN
-          expect(error.constraint).to.equal('organization_learners_userid_foreign');
-        });
-      });
-
-      it('throw one_active_learner', async function () {
-        const campaignParticipant = await makeCampaignParticipant({
-          campaignAttributes: { idPixLabel: null },
-          userIdentity,
-          participantExternalId: null,
-          isRestricted: false,
-        });
-
-        const error = await catchErr(() => {
-          return DomainTransaction.execute(async (domainTransaction) => {
-            await Promise.all([
-              campaignParticipantRepository.save(campaignParticipant, domainTransaction),
-              campaignParticipantRepository.save(campaignParticipant, domainTransaction),
-            ]);
-          });
-        })();
-
-        expect(error).to.be.instanceOf(OrganizationLearnersCouldNotBeSavedError);
-      });
-
-      it('does not update data', async function () {
-        const campaign = databaseBuilder.factory.buildCampaign({
-          multipleSendings: true,
-          idPixLabel: null,
-        });
-        const { id: previousCampaignParticipationForUserId } = databaseBuilder.factory.buildCampaignParticipation({
-          userId: userIdentity.id,
-          campaignId: campaign.id,
-          isImproved: false,
-        });
-
-        await databaseBuilder.commit();
-
-        const campaignToStartParticipation = new CampaignToStartParticipation({ ...campaign, id: 13 });
-        const campaignParticipant = new CampaignParticipant({
-          campaignToStartParticipation,
-          userIdentity,
-          previousCampaignParticipationForUser: {
-            id: previousCampaignParticipationForUserId,
-            status: 'SHARED',
-            validatedSkillsCount: 0,
-          },
-          organizationLearner: {
-            id: null,
-            hasParticipated: false,
-          },
-        });
-
-        campaignParticipant.start({ participantExternalId: null });
-
-        //WHEN
-        await catchErr(() => {
-          return DomainTransaction.execute(async (domainTransaction) => {
-            await campaignParticipantRepository.save(campaignParticipant, domainTransaction);
-          });
-        })();
-
-        //THEN
-        const campaignParticipations = await knex('campaign-participations').select(['id', 'isImproved']);
-        const assessments = await knex('assessments');
-
-        expect(campaignParticipations).to.deep.equal([
-          { id: previousCampaignParticipationForUserId, isImproved: false },
-        ]);
-        expect(assessments).to.be.empty;
-      });
-    });
-  });
-
   describe('get', function () {
     let organizationId;
     beforeEach(function () {
@@ -1102,54 +514,6 @@ describe('Integration | Infrastructure | Repository | CampaignParticipant', func
     });
   });
 });
-
-function getExpectedCampaignParticipation(campaignParticipationId, campaignParticipant) {
-  return {
-    ...pick(campaignParticipant.campaignParticipation, [
-      'id',
-      'campaignId',
-      'userId',
-      'status',
-      'organizationLearnerId',
-      'participantExternalId',
-    ]),
-    id: campaignParticipationId,
-  };
-}
-
-function getExpectedAssessment(campaignParticipationId, campaignParticipant) {
-  return {
-    ...pick(campaignParticipant.assessment, assessmentAttributes),
-    campaignParticipationId,
-  };
-}
-
-async function makeCampaignParticipant({
-  campaignAttributes,
-  userIdentity,
-  organizationLearnerId,
-  participantExternalId,
-  isRestricted,
-}) {
-  const campaign = databaseBuilder.factory.buildCampaign(campaignAttributes);
-
-  await databaseBuilder.commit();
-
-  const campaignToStartParticipation = new CampaignToStartParticipation({ ...campaign, isRestricted });
-  const organizationLearner = {
-    id: organizationLearnerId,
-    hasParticipated: false,
-  };
-  const campaignParticipant = new CampaignParticipant({
-    campaignToStartParticipation,
-    organizationLearner,
-    userIdentity,
-    previousCampaignParticipationForUser: null,
-  });
-
-  campaignParticipant.start({ participantExternalId });
-  return campaignParticipant;
-}
 
 function buildCampaignWithSkills(attributes, skills = ['skill1']) {
   const { id: organizationId } = databaseBuilder.factory.buildOrganization({
