+++ api/tests/integration/infrastructure/repositories/campaign-participation-repository_test.js
@@ -1,4 +1,4 @@
-import { sinon, expect, knex, databaseBuilder, catchErr } from '../../../test-helper.js';
+import { expect, knex, databaseBuilder, catchErr } from '../../../test-helper.js';
 import { Campaign } from '../../../../lib/domain/models/Campaign.js';
 import { Assessment } from '../../../../src/shared/domain/models/Assessment.js';
 import { CampaignParticipation } from '../../../../lib/domain/models/CampaignParticipation.js';
@@ -293,50 +293,6 @@ describe('Integration | Repository | Campaign Participation', function () {
     });
   });
 
-  describe('update', function () {
-    it('save the changes of the campaignParticipation', async function () {
-      const campaignParticipationId = 12;
-      const campaignParticipationToUpdate = databaseBuilder.factory.buildCampaignParticipation({
-        id: campaignParticipationId,
-        status: STARTED,
-        sharedAt: null,
-      });
-
-      await databaseBuilder.commit();
-
-      await campaignParticipationRepository.update({
-        ...campaignParticipationToUpdate,
-        sharedAt: new Date('2021-01-01'),
-        status: SHARED,
-      });
-      const campaignParticipation = await knex('campaign-participations')
-        .where({ id: campaignParticipationId })
-        .first();
-
-      expect(campaignParticipation.sharedAt).to.deep.equals(new Date('2021-01-01'));
-      expect(campaignParticipation.status).to.equals(SHARED);
-    });
-
-    it('should not update because the leaner can not have 2 active participations for the same campaign', async function () {
-      const campaignId = databaseBuilder.factory.buildCampaign().id;
-      const userId = databaseBuilder.factory.buildUser().id;
-      const organizationLearnerId = databaseBuilder.factory.buildOrganizationLearner({ userId }).id;
-      databaseBuilder.factory.buildCampaignParticipation({ organizationLearnerId, campaignId });
-      const campaignParticipationToUpdate = databaseBuilder.factory.buildCampaignParticipation({
-        organizationLearnerId,
-      });
-
-      await databaseBuilder.commit();
-
-      const error = await catchErr(campaignParticipationRepository.update)({
-        ...campaignParticipationToUpdate,
-        campaignId,
-      });
-
-      expect(error.constraint).to.equals('one_active_participation_by_learner');
-    });
-  });
-
   describe('#findLatestOngoingByUserId', function () {
     let userId;
 
@@ -496,94 +452,6 @@ describe('Integration | Repository | Campaign Participation', function () {
 
       // then
       expect(response).to.equal(null);
-    });
-  });
-
-  describe('#updateWithSnapshot', function () {
-    let clock;
-    let campaignParticipation;
-    const frozenTime = new Date('1987-09-01T00:00:00Z');
-
-    beforeEach(async function () {
-      campaignParticipation = databaseBuilder.factory.buildCampaignParticipation({
-        status: STARTED,
-        sharedAt: null,
-      });
-
-      databaseBuilder.factory.buildKnowledgeElement({
-        userId: campaignParticipation.userId,
-        createdAt: new Date('1985-09-01T00:00:00Z'),
-      });
-      clock = sinon.useFakeTimers({ now: frozenTime, toFake: ['Date'] });
-
-      await databaseBuilder.commit();
-    });
-
-    afterEach(function () {
-      clock.restore();
-    });
-
-    it('persists the campaign-participation changes', async function () {
-      // given
-      campaignParticipation.campaign = {};
-      campaignParticipation.assessments = [];
-      campaignParticipation.user = {};
-      campaignParticipation.assessmentId = {};
-      campaignParticipation.isShared = true;
-      campaignParticipation.status = SHARED;
-      campaignParticipation.participantExternalId = 'Laura';
-
-      // when
-      await campaignParticipationRepository.updateWithSnapshot(campaignParticipation);
-
-      const updatedCampaignParticipation = await knex('campaign-participations')
-        .where({ id: campaignParticipation.id })
-        .first();
-      // then
-      expect(updatedCampaignParticipation.status).to.equals(SHARED);
-      expect(updatedCampaignParticipation.participantExternalId).to.equals('Laura');
-    });
-
-    it('should save a snapshot', async function () {
-      // given
-      campaignParticipation.sharedAt = new Date();
-
-      // when
-      await campaignParticipationRepository.updateWithSnapshot(campaignParticipation);
-
-      // then
-      const snapshotInDB = await knex.select('id').from('knowledge-element-snapshots');
-      expect(snapshotInDB).to.have.length(1);
-    });
-
-    context('when there is a transaction', function () {
-      it('should save a snapshot using a transaction', async function () {
-        campaignParticipation.sharedAt = new Date();
-
-        await DomainTransaction.execute((domainTransaction) => {
-          return campaignParticipationRepository.updateWithSnapshot(campaignParticipation, domainTransaction);
-        });
-
-        const snapshotInDB = await knex.select('id').from('knowledge-element-snapshots');
-        expect(snapshotInDB).to.have.length(1);
-      });
-
-      it('does not save a snapshot when there is an error', async function () {
-        campaignParticipation.sharedAt = new Date();
-
-        try {
-          await DomainTransaction.execute(async (domainTransaction) => {
-            await campaignParticipationRepository.updateWithSnapshot(campaignParticipation, domainTransaction);
-            throw new Error();
-          });
-          // eslint-disable-next-line no-empty
-        } catch (error) {}
-
-        const snapshotInDB = await knex.select('id').from('knowledge-element-snapshots');
-        const participations = await knex.select('sharedAt').from('campaign-participations');
-        expect(participations.sharedAt).to.be.undefined;
-        expect(snapshotInDB).to.be.empty;
-      });
     });
   });
