+++ api/tests/unit/application/campaign-participations/index_test.js
@@ -4,50 +4,6 @@ import { campaignParticipationController } from '../../../../lib/application/cam
 import * as moduleUnderTest from '../../../../lib/application/campaign-participations/index.js';
 
 describe('Unit | Application | Router | campaign-participation-router ', function () {
-  describe('POST /api/campaign-participations', function () {
-    it('should return 200 when participant have external id', async function () {
-      // given
-      sinon.stub(campaignParticipationController, 'save').callsFake((request, h) => h.response('ok').code(200));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const participantExternalId = 'toto';
-      const response = await httpTestServer.request('POST', '/api/campaign-participations', {
-        data: {
-          type: 'campaign-participations',
-          attributes: {
-            'participant-external-id': participantExternalId,
-          },
-        },
-      });
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
-
-    it('should return 400 when participant external id exceeds 255 characters', async function () {
-      // given
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const participantExternalId256Characters =
-        '0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef';
-      const response = await httpTestServer.request('POST', '/api/campaign-participations', {
-        data: {
-          type: 'campaign-participations',
-          attributes: {
-            'participant-external-id': participantExternalId256Characters,
-          },
-        },
-      });
-
-      // then
-      expect(response.statusCode).to.equal(400);
-    });
-  });
-
   describe('PATCH /api/admin/campaign-participations/{id}', function () {
     it('returns 200 when admin member has rights', async function () {
       // given
